Resolve unnecessary_cast lints

+++ src/dumper.rs
@@ -15,15 +15,15 @@ pub unsafe fn yaml_emitter_open(mut emitter: *mut yaml_emitter_t) -> libc::c_int
     let mut event = MaybeUninit::<yaml_event_t>::uninit();
     let event = event.as_mut_ptr();
     let mark = yaml_mark_t {
-        index: 0 as libc::c_int as size_t,
-        line: 0 as libc::c_int as size_t,
-        column: 0 as libc::c_int as size_t,
+        index: 0_i32 as size_t,
+        line: 0_i32 as size_t,
+        column: 0_i32 as size_t,
     };
     __assert!(!emitter.is_null());
     __assert!((*emitter).opened == 0);
     memset(
         event as *mut libc::c_void,
-        0 as libc::c_int,
+        0_i32,
         size_of::<yaml_event_t>() as libc::c_ulong,
     );
     (*event).type_0 = YAML_STREAM_START_EVENT;
@@ -31,37 +31,37 @@ pub unsafe fn yaml_emitter_open(mut emitter: *mut yaml_emitter_t) -> libc::c_int
     (*event).end_mark = mark;
     (*event).data.stream_start.encoding = YAML_ANY_ENCODING;
     if yaml_emitter_emit(emitter, event) == 0 {
-        return 0 as libc::c_int;
+        return 0_i32;
     }
-    (*emitter).opened = 1 as libc::c_int;
-    1 as libc::c_int
+    (*emitter).opened = 1_i32;
+    1_i32
 }
 pub unsafe fn yaml_emitter_close(mut emitter: *mut yaml_emitter_t) -> libc::c_int {
     let mut event = MaybeUninit::<yaml_event_t>::uninit();
     let event = event.as_mut_ptr();
     let mark = yaml_mark_t {
-        index: 0 as libc::c_int as size_t,
-        line: 0 as libc::c_int as size_t,
-        column: 0 as libc::c_int as size_t,
+        index: 0_i32 as size_t,
+        line: 0_i32 as size_t,
+        column: 0_i32 as size_t,
     };
     __assert!(!emitter.is_null());
     __assert!((*emitter).opened != 0);
     if (*emitter).closed != 0 {
-        return 1 as libc::c_int;
+        return 1_i32;
     }
     memset(
         event as *mut libc::c_void,
-        0 as libc::c_int,
+        0_i32,
         size_of::<yaml_event_t>() as libc::c_ulong,
     );
     (*event).type_0 = YAML_STREAM_END_EVENT;
     (*event).start_mark = mark;
     (*event).end_mark = mark;
     if yaml_emitter_emit(emitter, event) == 0 {
-        return 0 as libc::c_int;
+        return 0_i32;
     }
-    (*emitter).closed = 1 as libc::c_int;
-    1 as libc::c_int
+    (*emitter).closed = 1_i32;
+    1_i32
 }
 pub unsafe fn yaml_emitter_dump(
     emitter: *mut yaml_emitter_t,
@@ -71,9 +71,9 @@ pub unsafe fn yaml_emitter_dump(
     let mut event = MaybeUninit::<yaml_event_t>::uninit();
     let event = event.as_mut_ptr();
     let mark = yaml_mark_t {
-        index: 0 as libc::c_int as size_t,
-        line: 0 as libc::c_int as size_t,
-        column: 0 as libc::c_int as size_t,
+        index: 0_i32 as size_t,
+        line: 0_i32 as size_t,
+        column: 0_i32 as size_t,
     };
     __assert!(!emitter.is_null());
     __assert!(!document.is_null());
@@ -93,7 +93,7 @@ pub unsafe fn yaml_emitter_dump(
             if (*document).nodes.start == (*document).nodes.top {
                 if !(yaml_emitter_close(emitter) == 0) {
                     yaml_emitter_delete_document_and_anchors(emitter);
-                    return 1 as libc::c_int;
+                    return 1_i32;
                 }
             } else {
                 __assert!((*emitter).opened != 0);
@@ -105,7 +105,7 @@ pub unsafe fn yaml_emitter_dump(
                 if !((*emitter).anchors).is_null() {
                     memset(
                         (*emitter).anchors as *mut libc::c_void,
-                        0 as libc::c_int,
+                        0_i32,
                         (size_of::<yaml_anchors_t>() as libc::c_ulong).wrapping_mul(
                             ((*document).nodes.top).c_offset_from((*document).nodes.start)
                                 as libc::c_long as libc::c_ulong,
@@ -113,7 +113,7 @@ pub unsafe fn yaml_emitter_dump(
                     );
                     memset(
                         event as *mut libc::c_void,
-                        0 as libc::c_int,
+                        0_i32,
                         size_of::<yaml_event_t>() as libc::c_ulong,
                     );
                     (*event).type_0 = YAML_DOCUMENT_START_EVENT;
@@ -126,11 +126,11 @@ pub unsafe fn yaml_emitter_dump(
                         (*document).tag_directives.end;
                     (*event).data.document_start.implicit = (*document).start_implicit;
                     if !(yaml_emitter_emit(emitter, event) == 0) {
-                        yaml_emitter_anchor_node(emitter, 1 as libc::c_int);
-                        if !(yaml_emitter_dump_node(emitter, 1 as libc::c_int) == 0) {
+                        yaml_emitter_anchor_node(emitter, 1_i32);
+                        if !(yaml_emitter_dump_node(emitter, 1_i32) == 0) {
                             memset(
                                 event as *mut libc::c_void,
-                                0 as libc::c_int,
+                                0_i32,
                                 size_of::<yaml_event_t>() as libc::c_ulong,
                             );
                             (*event).type_0 = YAML_DOCUMENT_END_EVENT;
@@ -139,7 +139,7 @@ pub unsafe fn yaml_emitter_dump(
                             (*event).data.document_end.implicit = (*document).end_implicit;
                             if !(yaml_emitter_emit(emitter, event) == 0) {
                                 yaml_emitter_delete_document_and_anchors(emitter);
-                                return 1 as libc::c_int;
+                                return 1_i32;
                             }
                         }
                     }
@@ -149,7 +149,7 @@ pub unsafe fn yaml_emitter_dump(
         _ => {}
     }
     yaml_emitter_delete_document_and_anchors(emitter);
-    0 as libc::c_int
+    0_i32
 }
 unsafe fn yaml_emitter_delete_document_and_anchors(mut emitter: *mut yaml_emitter_t) {
     let mut index: libc::c_int;
@@ -159,7 +159,7 @@ unsafe fn yaml_emitter_delete_document_and_anchors(mut emitter: *mut yaml_emitte
         *fresh2 = ptr::null_mut::<yaml_document_t>();
         return;
     }
-    index = 0 as libc::c_int;
+    index = 0_i32;
     while ((*(*emitter).document).nodes.start).wrapping_offset(index as isize)
         < (*(*emitter).document).nodes.top
     {
@@ -195,23 +195,20 @@ unsafe fn yaml_emitter_delete_document_and_anchors(mut emitter: *mut yaml_emitte
     yaml_free((*emitter).anchors as *mut libc::c_void);
     let fresh6 = addr_of_mut!((*emitter).anchors);
     *fresh6 = ptr::null_mut::<yaml_anchors_t>();
-    (*emitter).last_anchor_id = 0 as libc::c_int;
+    (*emitter).last_anchor_id = 0_i32;
     let fresh7 = addr_of_mut!((*emitter).document);
     *fresh7 = ptr::null_mut::<yaml_document_t>();
 }
 unsafe fn yaml_emitter_anchor_node(emitter: *mut yaml_emitter_t, index: libc::c_int) {
     let node: *mut yaml_node_t = ((*(*emitter).document).nodes.start)
         .wrapping_offset(index as isize)
-        .wrapping_offset(-(1 as libc::c_int as isize));
+        .wrapping_offset(-(1_i32 as isize));
     let mut item: *mut yaml_node_item_t;
     let mut pair: *mut yaml_node_pair_t;
-    let fresh8 = addr_of_mut!(
-        (*((*emitter).anchors).wrapping_offset((index - 1 as libc::c_int) as isize)).references
-    );
+    let fresh8 =
+        addr_of_mut!((*((*emitter).anchors).wrapping_offset((index - 1_i32) as isize)).references);
     *fresh8 += 1;
-    if (*((*emitter).anchors).wrapping_offset((index - 1 as libc::c_int) as isize)).references
-        == 1 as libc::c_int
-    {
+    if (*((*emitter).anchors).wrapping_offset((index - 1_i32) as isize)).references == 1_i32 {
         match (*node).type_0 as libc::c_uint {
             2 => {
                 item = (*node).data.sequence.items.start;
@@ -230,21 +227,18 @@ unsafe fn yaml_emitter_anchor_node(emitter: *mut yaml_emitter_t, index: libc::c_
             }
             _ => {}
         }
-    } else if (*((*emitter).anchors).wrapping_offset((index - 1 as libc::c_int) as isize))
-        .references
-        == 2 as libc::c_int
+    } else if (*((*emitter).anchors).wrapping_offset((index - 1_i32) as isize)).references == 2_i32
     {
         let fresh9 = addr_of_mut!((*emitter).last_anchor_id);
         *fresh9 += 1;
-        (*((*emitter).anchors).wrapping_offset((index - 1 as libc::c_int) as isize)).anchor =
-            *fresh9;
+        (*((*emitter).anchors).wrapping_offset((index - 1_i32) as isize)).anchor = *fresh9;
     }
 }
 unsafe fn yaml_emitter_generate_anchor(
     _emitter: *mut yaml_emitter_t,
     anchor_id: libc::c_int,
 ) -> *mut yaml_char_t {
-    let anchor: *mut yaml_char_t = yaml_malloc(16 as libc::c_int as size_t) as *mut yaml_char_t;
+    let anchor: *mut yaml_char_t = yaml_malloc(16_i32 as size_t) as *mut yaml_char_t;
     if anchor.is_null() {
         return ptr::null_mut::<yaml_char_t>();
     }
@@ -255,22 +249,20 @@ unsafe fn yaml_emitter_generate_anchor(
 unsafe fn yaml_emitter_dump_node(emitter: *mut yaml_emitter_t, index: libc::c_int) -> libc::c_int {
     let node: *mut yaml_node_t = ((*(*emitter).document).nodes.start)
         .wrapping_offset(index as isize)
-        .wrapping_offset(-(1 as libc::c_int as isize));
+        .wrapping_offset(-(1_i32 as isize));
     let anchor_id: libc::c_int =
-        (*((*emitter).anchors).wrapping_offset((index - 1 as libc::c_int) as isize)).anchor;
+        (*((*emitter).anchors).wrapping_offset((index - 1_i32) as isize)).anchor;
     let mut anchor: *mut yaml_char_t = ptr::null_mut::<yaml_char_t>();
     if anchor_id != 0 {
         anchor = yaml_emitter_generate_anchor(emitter, anchor_id);
         if anchor.is_null() {
-            return 0 as libc::c_int;
+            return 0_i32;
         }
     }
-    if (*((*emitter).anchors).wrapping_offset((index - 1 as libc::c_int) as isize)).serialized != 0
-    {
+    if (*((*emitter).anchors).wrapping_offset((index - 1_i32) as isize)).serialized != 0 {
         return yaml_emitter_dump_alias(emitter, anchor);
     }
-    (*((*emitter).anchors).wrapping_offset((index - 1 as libc::c_int) as isize)).serialized =
-        1 as libc::c_int;
+    (*((*emitter).anchors).wrapping_offset((index - 1_i32) as isize)).serialized = 1_i32;
     match (*node).type_0 as libc::c_uint {
         1 => yaml_emitter_dump_scalar(emitter, node, anchor),
         2 => yaml_emitter_dump_sequence(emitter, node, anchor),
@@ -285,13 +277,13 @@ unsafe fn yaml_emitter_dump_alias(
     let mut event = MaybeUninit::<yaml_event_t>::uninit();
     let event = event.as_mut_ptr();
     let mark = yaml_mark_t {
-        index: 0 as libc::c_int as size_t,
-        line: 0 as libc::c_int as size_t,
-        column: 0 as libc::c_int as size_t,
+        index: 0_i32 as size_t,
+        line: 0_i32 as size_t,
+        column: 0_i32 as size_t,
     };
     memset(
         event as *mut libc::c_void,
-        0 as libc::c_int,
+        0_i32,
         size_of::<yaml_event_t>() as libc::c_ulong,
     );
     (*event).type_0 = YAML_ALIAS_EVENT;
@@ -308,21 +300,21 @@ unsafe fn yaml_emitter_dump_scalar(
     let mut event = MaybeUninit::<yaml_event_t>::uninit();
     let event = event.as_mut_ptr();
     let mark = yaml_mark_t {
-        index: 0 as libc::c_int as size_t,
-        line: 0 as libc::c_int as size_t,
-        column: 0 as libc::c_int as size_t,
+        index: 0_i32 as size_t,
+        line: 0_i32 as size_t,
+        column: 0_i32 as size_t,
     };
     let plain_implicit: libc::c_int = (strcmp(
         (*node).tag as *mut libc::c_char,
         b"tag:yaml.org,2002:str\0" as *const u8 as *const libc::c_char,
-    ) == 0 as libc::c_int) as libc::c_int;
+    ) == 0_i32) as libc::c_int;
     let quoted_implicit: libc::c_int = (strcmp(
         (*node).tag as *mut libc::c_char,
         b"tag:yaml.org,2002:str\0" as *const u8 as *const libc::c_char,
-    ) == 0 as libc::c_int) as libc::c_int;
+    ) == 0_i32) as libc::c_int;
     memset(
         event as *mut libc::c_void,
-        0 as libc::c_int,
+        0_i32,
         size_of::<yaml_event_t>() as libc::c_ulong,
     );
     (*event).type_0 = YAML_SCALAR_EVENT;
@@ -345,18 +337,18 @@ unsafe fn yaml_emitter_dump_sequence(
     let mut event = MaybeUninit::<yaml_event_t>::uninit();
     let event = event.as_mut_ptr();
     let mark = yaml_mark_t {
-        index: 0 as libc::c_int as size_t,
-        line: 0 as libc::c_int as size_t,
-        column: 0 as libc::c_int as size_t,
+        index: 0_i32 as size_t,
+        line: 0_i32 as size_t,
+        column: 0_i32 as size_t,
     };
     let implicit: libc::c_int = (strcmp(
         (*node).tag as *mut libc::c_char,
         b"tag:yaml.org,2002:seq\0" as *const u8 as *const libc::c_char,
-    ) == 0 as libc::c_int) as libc::c_int;
+    ) == 0_i32) as libc::c_int;
     let mut item: *mut yaml_node_item_t;
     memset(
         event as *mut libc::c_void,
-        0 as libc::c_int,
+        0_i32,
         size_of::<yaml_event_t>() as libc::c_ulong,
     );
     (*event).type_0 = YAML_SEQUENCE_START_EVENT;
@@ -367,27 +359,27 @@ unsafe fn yaml_emitter_dump_sequence(
     (*event).data.sequence_start.implicit = implicit;
     (*event).data.sequence_start.style = (*node).data.sequence.style;
     if yaml_emitter_emit(emitter, event) == 0 {
-        return 0 as libc::c_int;
+        return 0_i32;
     }
     item = (*node).data.sequence.items.start;
     while item < (*node).data.sequence.items.top {
         if yaml_emitter_dump_node(emitter, *item) == 0 {
-            return 0 as libc::c_int;
+            return 0_i32;
         }
         item = item.wrapping_offset(1);
     }
     memset(
         event as *mut libc::c_void,
-        0 as libc::c_int,
+        0_i32,
         size_of::<yaml_event_t>() as libc::c_ulong,
     );
     (*event).type_0 = YAML_SEQUENCE_END_EVENT;
     (*event).start_mark = mark;
     (*event).end_mark = mark;
     if yaml_emitter_emit(emitter, event) == 0 {
-        return 0 as libc::c_int;
+        return 0_i32;
     }
-    1 as libc::c_int
+    1_i32
 }
 unsafe fn yaml_emitter_dump_mapping(
     emitter: *mut yaml_emitter_t,
@@ -397,18 +389,18 @@ unsafe fn yaml_emitter_dump_mapping(
     let mut event = MaybeUninit::<yaml_event_t>::uninit();
     let event = event.as_mut_ptr();
     let mark = yaml_mark_t {
-        index: 0 as libc::c_int as size_t,
-        line: 0 as libc::c_int as size_t,
-        column: 0 as libc::c_int as size_t,
+        index: 0_i32 as size_t,
+        line: 0_i32 as size_t,
+        column: 0_i32 as size_t,
     };
     let implicit: libc::c_int = (strcmp(
         (*node).tag as *mut libc::c_char,
         b"tag:yaml.org,2002:map\0" as *const u8 as *const libc::c_char,
-    ) == 0 as libc::c_int) as libc::c_int;
+    ) == 0_i32) as libc::c_int;
     let mut pair: *mut yaml_node_pair_t;
     memset(
         event as *mut libc::c_void,
-        0 as libc::c_int,
+        0_i32,
         size_of::<yaml_event_t>() as libc::c_ulong,
     );
     (*event).type_0 = YAML_MAPPING_START_EVENT;
@@ -419,28 +411,28 @@ unsafe fn yaml_emitter_dump_mapping(
     (*event).data.mapping_start.implicit = implicit;
     (*event).data.mapping_start.style = (*node).data.mapping.style;
     if yaml_emitter_emit(emitter, event) == 0 {
-        return 0 as libc::c_int;
+        return 0_i32;
     }
     pair = (*node).data.mapping.pairs.start;
     while pair < (*node).data.mapping.pairs.top {
         if yaml_emitter_dump_node(emitter, (*pair).key) == 0 {
-            return 0 as libc::c_int;
+            return 0_i32;
         }
         if yaml_emitter_dump_node(emitter, (*pair).value) == 0 {
-            return 0 as libc::c_int;
+            return 0_i32;
         }
         pair = pair.wrapping_offset(1);
     }
     memset(
         event as *mut libc::c_void,
-        0 as libc::c_int,
+        0_i32,
         size_of::<yaml_event_t>() as libc::c_ulong,
     );
     (*event).type_0 = YAML_MAPPING_END_EVENT;
     (*event).start_mark = mark;
     (*event).end_mark = mark;
     if yaml_emitter_emit(emitter, event) == 0 {
-        return 0 as libc::c_int;
+        return 0_i32;
     }
-    1 as libc::c_int
+    1_i32
 }
